feat: Inicio da implementação de iA é preciso refatortar e melhorar
