--- src/main/java/org/back/domain/repository/AlternativaRepository.java
@@ -1,10 +1,0 @@
-package org.back.domain.repository;
-
-import org.back.domain.missoes.Alternativa;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AlternativaRepository extends JpaRepository<Alternativa, Integer> {
-
-}
