--- src/main/java/org/back/domain/repository/MesaRepository.java
@@ -1,10 +1,0 @@
-package org.back.domain.repository;
-
-import org.back.domain.mesa.Mesa;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MesaRepository extends JpaRepository<Mesa, Integer> {
-
-}
