--- src/main/java/org/back/domain/repository/MissaoRepository.java
@@ -1,10 +1,0 @@
-package org.back.domain.repository;
-
-import org.back.domain.missoes.Missao;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MissaoRepository extends JpaRepository<Missao, Integer> {
-
-}
