--- src/main/java/org/back/domain/repository/QuestaoRepository.java
@@ -1,10 +1,0 @@
-package org.back.domain.repository;
-
-import org.back.domain.missoes.Questao;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuestaoRepository extends JpaRepository<Questao, Integer> {
-
-}
