--- src/main/java/org/back/domain/repository/UsuarioRepository.java
@@ -1,14 +1,0 @@
-package org.back.domain.repository;
-
-import org.back.domain.usuario.Usuario;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UsuarioRepository  extends JpaRepository<Usuario, Integer> {
-
-    Optional<Usuario> findByMatricula(String matricula);
-
-}
