+++ src/main/java/org/back/src/conversor/MapperConvert.java
@@ -1,0 +1,11 @@
+package org.back.src.conversor;
+
+import org.modelmapper.ModelMapper;
+
+public class MapperConvert {
+    private static final ModelMapper modelMapper = new ModelMapper();
+
+    public static <D, T> D convert(T entity, Class<D> newEntityClass) {
+        return modelMapper.map(entity, newEntityClass);
+    }
+}
