+++ src/main/java/org/back/src/enums/ETipoUsuario.java
@@ -1,0 +1,19 @@
+package org.back.src.enums;
+
+public enum ETipoUsuario {
+
+    JOGADOR(0),
+    MESTRE(1);
+
+    private final int valor;
+
+    ETipoUsuario(int valor) {
+        this.valor = valor;
+    }
+
+    public int getValor() {
+        return valor;
+    }
+
+}
+
