+++ src/main/java/org/back/src/exception/CustomException.java
@@ -1,0 +1,33 @@
+package org.back.src.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class CustomException extends Exception {
+
+    private HttpStatus status;
+    private String message;
+
+    public CustomException() {}
+
+    public CustomException(HttpStatus status, String message) {
+        this.status = status;
+        this.message = message;
+    }
+
+    public HttpStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(HttpStatus status) {
+        this.status = status;
+    }
+
+    @Override
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
