+++ src/main/java/org/back/src/repository/AlternativaRepository.java
@@ -1,0 +1,10 @@
+package org.back.src.repository;
+
+import org.back.src.entity.missoes.Alternativa;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AlternativaRepository extends JpaRepository<Alternativa, Integer> {
+
+}
