+++ src/main/java/org/back/src/repository/MesaRepository.java
@@ -1,0 +1,10 @@
+package org.back.src.repository;
+
+import org.back.src.entity.mesa.Mesa;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MesaRepository extends JpaRepository<Mesa, Integer> {
+
+}
