+++ src/main/java/org/back/src/service/UsuarioDetailsService.java
@@ -1,0 +1,24 @@
+package org.back.src.service;
+
+import org.back.src.repository.UsuarioRepository;
+import org.back.src.entity.usuario.Usuario;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UsuarioDetailsService implements UserDetailsService {
+
+    @Autowired
+    private UsuarioRepository usuarioRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String matricula) throws UsernameNotFoundException {
+        Usuario usuario = usuarioRepository.findByMatricula(matricula)
+                .orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado: " + matricula));
+        return usuario;
+    }
+
+}
